Add source of the sbt-plugin
Originally from https://github.com/scalacenter/sbt-github-dependency-graph
I think it is better to have the sbt-plugin and the Github action
in the same place because the sbt-plugin is a implementation detail
of the Github action. The sbt-plugin cannot be used outside the Github action.
The two are tightly coupled and must be released together.